update IAction.cs code and change IDamageable target to Character target

+++ TextBasedRPG_/Model/Character.cs
@@ -16,13 +16,13 @@ public abstract class Character : IDamageable, IIDentifiable
     public required string Id { get; set; }
     public ElementType Element { get; set; }
     
-    public int Health { get; set; }
-    public int MaxHealth { get; set; }
-    public int AttackPower { get; set; }
+    public double Health { get; set; }
+    public double MaxHealth { get; set; }
+    public double AttackPower { get; set; }
 
     // public Character(string Name, int AttackPower)
     // {}
 
-    public abstract void TakeDamage(int damage);
+    public abstract void TakeDamage(double damage);
 
 }
