+++ TextBasedRPG_/Model/Heroes.cs
@@ -12,7 +12,7 @@ public class Heroes : Character
         AttackPower = attackPower;
         Element = element;
     }
-    public override void TakeDamage(int damage)
+    public override void TakeDamage(double damage)
     {
         Health -= damage;
         Console.WriteLine($"{Name} has taken {damage} damage, and now has {Health} health.");
